refactor(api): ♻️ 引入 MapStruct 并统一 DTO/Entity 转换逻辑
- 新增 ClusterServiceRoleInstanceWebuisConverter、ClusterServiceCommandHostCommandConverter、NoticeGroupConverter 等 MapStruct Mapper，替代手写 set/get
- ClusterServiceRoleInstanceWebuisEntity 主键类型由 Integer 调整为 Long，避免主键溢出风险
- Entity/DTO/VO 新增 serviceName 字段，用于前端统一显示服务图标
- ServiceInstallation/CommandExecution/ServiceCommand 等业务流程统一使用 MapStruct 进行对象转换，减少样板代码
- UI 侧 ClusterPage 列表中的服务图标渲染逻辑改为按 serviceName 取 SVG，保证图标一致性
- 批量填充 ClusterServiceCommandHostCommandEntity.serviceName，避免 N+1 查询

+++ datasophon-dao/src/main/java/com/datasophon/dao/entity/ClusterServiceRoleInstanceWebuisEntity.java
@@ -48,15 +48,18 @@ public class ClusterServiceRoleInstanceWebuisEntity extends BaseEntity {
 
 
     /**
-     * 服务角色id
+     * 服务角色实例ID（外键）
      */
-    private Integer serviceRoleInstanceId;
+    private Long serviceRoleInstanceId;
     /**
      * URL地址
      */
     private String webUrl;
 
-    private Integer serviceInstanceId;
+    /**
+     * 服务实例ID（外键）
+     */
+    private Long serviceInstanceId;
 
     private String name;
